Add files via upload
It is a general design of how our mapping system will work. Instead of getting random values from random.py library, we will take the data from the Arduino prompt, that are going to be exchanged with the CanSat via LoRa, and then map them on screen. We will map the CanSat's location, as well as the location of the people found